requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
cryptography==42.0.8
python-dotenv==1.0.1
Werkzeug==3.0.3
